Add files via upload
A fun and interactive game where players are tasked with comparing the number of social media followers of two random figures or brands, ranging from celebrities to organizations. The goal is to correctly guess which one has more followers, earning points for every correct answer. The game ends when the player makes an incorrect guess, and their final score is displayed.